Support for the Kegger in the Woods
git-svn-id: http://svn.code.sf.net/p/kolmafia/code@8441 b29ace70-8910-0410-8dcc-aa2fc6433167

+++ src/net/sourceforge/kolmafia/session/ValhallaManager.java
@@ -233,7 +233,7 @@ public class ValhallaManager
 
 	public static final void resetPerAscensionCounters()
 	{
-		Preferences.setFloat( "slimelingFullness", (float) 0.0 );
+		Preferences.setFloat( "slimelingFullness", 0.0F );
 		Preferences.setInteger( "currentBountyItem", 0 );
 		Preferences.setString( "currentHippyStore", "none" );
 		Preferences.setString( "currentWheelPosition", "muscle" );
